moved the swagger setup to the app.js to load that before the react app is loaded.
