+++ server/app.js
@@ -1,6 +1,8 @@
 const cors = require('cors')
 const logger = require('./logger')
 const cookieParser = require('cookie-parser')
+const swaggerInline = require('swagger-inline')
+const swaggerUi = require('swagger-ui-express')
 
 require('dotenv').config()
 
@@ -39,61 +41,85 @@ app.use(logger.expressLogger)
 app.use(cors())
 app.use(express.json()) //req.body
 
-app.use('/auth', authentication)
+swaggerInline(['./*.js', './routes/*.js'], {
+  base: 'swaggerBase.json',
+}).then((generatedSwagger) => {
+  const swaggerDocument = JSON.parse(generatedSwagger)
 
-// Using the auth checks is undocumented, so disabled by default
-if (process.env.ENABLE_AUTH) {
-  console.log('Authentication is ENABLED')
-  app.use('/api/v1/categories', authentic, authorize, categories)
+  const options = {
+    jsonEditor: true,
+  }
   app.use(
-    '/api/v1/geospatialDefinitions',
-    authentic,
-    authorize,
-    geospatialDefinitions
+    '/api-docs',
+    function (req, res, next) {
+      swaggerDocument.host = req.get('host')
+      req.swaggerDoc = swaggerDocument
+      next()
+    },
+    swaggerUi.serve,
+    swaggerUi.setup(null, options)
   )
-  app.use('/api/v1/officials', authentic, authorize, officials)
-  app.use('/api/v1/channels', authentic, authorize, channels)
-  app.use('/api/v1/videos', authentic, authorize, videos)
-  app.use('/api/v1/advocacyGroups', authentic, authorize, advocacyGroups)
-  app.use('/api/v1/publications', authentic, authorize, publications)
-  app.use('/api/v1/videoTestimonials', authentic, authorize, videoTestimonials)
-  app.use('/api/v1/levels', authentic, authorize, levels)
-  app.use(
-    '/api/v1/legislativeArtifacts',
-    authentic,
-    authorize,
-    legislativeArtifacts
-  )
-  app.use(
-    '/api/v1/adminIntersections',
-    authentic,
-    authorize,
-    adminIntersections
-  )
-} else {
-  console.log('Authentication is DISABLED')
-  app.use('/api/v1/categories', categories)
-  app.use('/api/v1/geospatialDefinitions', geospatialDefinitions)
-  app.use('/api/v1/officials', officials)
-  app.use('/api/v1/channels', channels)
-  app.use('/api/v1/videos', videos)
-  app.use('/api/v1/advocacyGroups', advocacyGroups)
-  app.use('/api/v1/publications', publications)
-  app.use('/api/v1/videoTestimonials', videoTestimonials)
-  app.use('/api/v1/levels', levels)
-  app.use('/api/v1/legislativeArtifacts', legislativeArtifacts)
-  app.use('/api/v1/adminIntersections', adminIntersections)
-}
+  app.use('/auth', authentication)
 
-app.use(
-  express.static(path.join(__dirname, './admin/build'), {
-    fallthrough: true,
-  })
-)
+  // Using the auth checks is undocumented, so disabled by default
+  if (process.env.ENABLE_AUTH) {
+    console.log('Authentication is ENABLED')
+    app.use('/api/v1/categories', authentic, authorize, categories)
+    app.use(
+      '/api/v1/geospatialDefinitions',
+      authentic,
+      authorize,
+      geospatialDefinitions
+    )
+    app.use('/api/v1/officials', authentic, authorize, officials)
+    app.use('/api/v1/channels', authentic, authorize, channels)
+    app.use('/api/v1/videos', authentic, authorize, videos)
+    app.use('/api/v1/advocacyGroups', authentic, authorize, advocacyGroups)
+    app.use('/api/v1/publications', authentic, authorize, publications)
+    app.use(
+      '/api/v1/videoTestimonials',
+      authentic,
+      authorize,
+      videoTestimonials
+    )
+    app.use('/api/v1/levels', authentic, authorize, levels)
+    app.use(
+      '/api/v1/legislativeArtifacts',
+      authentic,
+      authorize,
+      legislativeArtifacts
+    )
+    app.use(
+      '/api/v1/adminIntersections',
+      authentic,
+      authorize,
+      adminIntersections
+    )
+  } else {
+    console.log('Authentication is DISABLED')
+    app.use('/api/v1/categories', categories)
+    app.use('/api/v1/geospatialDefinitions', geospatialDefinitions)
+    app.use('/api/v1/officials', officials)
+    app.use('/api/v1/channels', channels)
+    app.use('/api/v1/videos', videos)
+    app.use('/api/v1/advocacyGroups', advocacyGroups)
+    app.use('/api/v1/publications', publications)
+    app.use('/api/v1/videoTestimonials', videoTestimonials)
+    app.use('/api/v1/levels', levels)
+    app.use('/api/v1/legislativeArtifacts', legislativeArtifacts)
+    app.use('/api/v1/adminIntersections', adminIntersections)
+  }
 
-app.get('*', function (req, res) {
-  res.sendFile('index.html', {
-    root: path.join(__dirname, './admin/build/'),
+  app.use(
+    express.static(path.join(__dirname, './admin/build'), {
+      fallthrough: true,
+    })
+  )
+
+  app.get('*', function (req, res) {
+    res.sendFile('index.html', {
+      root: path.join(__dirname, './admin/build/'),
+    })
   })
 })
